리팩토링: 프로젝트 페이지 내의 헤더 컴포넌트화
프로젝트페이지 내의 헤더를 컴포넌트화하여 분리하고 리뷰내용을 바탕으로 리팩토링하였습니다.
